chore: fix some mistypo
